Move Cas2Application integration tests to own test file
