+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/integration/ApplicationTest.kt
@@ -21,8 +21,6 @@ import org.springframework.test.web.reactive.server.returnResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Application
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ApprovedPremisesApplication
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ApprovedPremisesApplicationSummary
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Cas2Application
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Cas2ApplicationSummary
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.FullPerson
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.NewApplication
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.NewWithdrawal
@@ -30,13 +28,11 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.OfflineApplica
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ReleaseTypeOption
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ServiceName
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.SubmitApprovedPremisesApplication
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.SubmitCas2Application
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.SubmitTemporaryAccommodationApplication
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.TemporaryAccommodationApplication
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.TemporaryAccommodationApplicationSummary
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.UpdateApplicationType
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.UpdateApprovedPremisesApplication
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.UpdateCas2Application
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.WithdrawalReason
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.NeedsDetailsFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.OffenderDetailsSummaryFactory
@@ -108,64 +104,6 @@ class ApplicationTest : IntegrationTestBase() {
       .exchange()
       .expectStatus()
       .isUnauthorized
-  }
-
-  @Test
-  fun `Get all applications returns 200 with correct body - when the service is CAS2`() {
-    `Given a User` { userEntity, jwt ->
-      `Given a User` { otherUser, _ ->
-        `Given an Offender` { offenderDetails, _ ->
-          cas2ApplicationJsonSchemaRepository.deleteAll()
-
-          val applicationSchema = cas2ApplicationJsonSchemaEntityFactory.produceAndPersist {
-            withAddedAt(OffsetDateTime.now())
-            withId(UUID.randomUUID())
-          }
-
-          val cas2ApplicationEntity = cas2ApplicationEntityFactory.produceAndPersist {
-            withApplicationSchema(applicationSchema)
-            withCreatedByUser(userEntity)
-            withCrn(offenderDetails.otherIds.crn)
-            withData("{}")
-          }
-
-          val otherCas2ApplicationEntity = cas2ApplicationEntityFactory.produceAndPersist {
-            withApplicationSchema(applicationSchema)
-            withCreatedByUser(otherUser)
-            withCrn(offenderDetails.otherIds.crn)
-            withData("{}")
-          }
-
-          CommunityAPI_mockOffenderUserAccessCall(userEntity.deliusUsername, offenderDetails.otherIds.crn, false, false)
-
-          val rawResponseBody = webTestClient.get()
-            .uri("/cas2/applications")
-            .header("Authorization", "Bearer $jwt")
-            .header("X-Service-Name", ServiceName.cas2.value)
-            .exchange()
-            .expectStatus()
-            .isOk
-            .returnResult<String>()
-            .responseBody
-            .blockFirst()
-
-          val responseBody =
-            objectMapper.readValue(rawResponseBody, object : TypeReference<List<Cas2ApplicationSummary>>() {})
-
-          assertThat(responseBody).anyMatch {
-            cas2ApplicationEntity.id == it.id &&
-              cas2ApplicationEntity.crn == it.person.crn &&
-              cas2ApplicationEntity.createdAt.toInstant() == it.createdAt &&
-              cas2ApplicationEntity.createdByUser.id == it.createdByUserId &&
-              cas2ApplicationEntity.submittedAt?.toInstant() == it.submittedAt
-          }
-
-          assertThat(responseBody).noneMatch {
-            otherCas2ApplicationEntity.id == it.id
-          }
-        }
-      }
-    }
   }
 
   @Test
@@ -1416,41 +1354,6 @@ class ApplicationTest : IntegrationTestBase() {
   }
 
   @Test
-  fun `Create new application for CAS-2 returns 201 with correct body and Location header`() {
-    `Given a User` { userEntity, jwt ->
-      `Given an Offender` { offenderDetails, _ ->
-        val applicationSchema = cas2ApplicationJsonSchemaEntityFactory.produceAndPersist {
-          withAddedAt(OffsetDateTime.now())
-          withId(UUID.randomUUID())
-        }
-
-        val result = webTestClient.post()
-          .uri("/cas2/applications")
-          .header("Authorization", "Bearer $jwt")
-          .header("X-Service-Name", ServiceName.cas2.value)
-          .bodyValue(
-            NewApplication(
-              crn = offenderDetails.otherIds.crn,
-            ),
-          )
-          .exchange()
-          .expectStatus()
-          .isCreated
-          .returnResult(Cas2Application::class.java)
-
-        assertThat(result.responseHeaders["Location"]).anyMatch {
-          it.matches(Regex("/applications/.+"))
-        }
-
-        assertThat(result.responseBody.blockFirst()).matches {
-          it.person.crn == offenderDetails.otherIds.crn &&
-            it.schemaVersion == applicationSchema.id
-        }
-      }
-    }
-  }
-
-  @Test
   fun `Update existing AP application returns 200 with correct body`() {
     `Given a User` { submittingUser, jwt ->
       `Given a User`(roles = listOf(UserRole.CAS1_ASSESSOR), qualifications = listOf(UserQualification.PIPE)) { assessorUser, _ ->
@@ -1509,65 +1412,6 @@ class ApplicationTest : IntegrationTestBase() {
 
           assertThat(result.person.crn).isEqualTo(offenderDetails.otherIds.crn)
         }
-      }
-    }
-  }
-
-  @Test
-  fun `Update existing CAS2 application returns 200 with correct body`() {
-    `Given a User` { submittingUser, jwt ->
-      `Given an Offender` { offenderDetails, _ ->
-        val applicationId = UUID.fromString("22ceda56-98b2-411d-91cc-ace0ab8be872")
-
-        val applicationSchema = cas2ApplicationJsonSchemaEntityFactory.produceAndPersist {
-          withAddedAt(OffsetDateTime.now())
-          withId(UUID.randomUUID())
-          withSchema(
-            """
-              {
-                "${"\$schema"}": "https://json-schema.org/draft/2020-12/schema",
-                "${"\$id"}": "https://example.com/product.schema.json",
-                "title": "Thing",
-                "description": "A thing",
-                "type": "object",
-                "properties": {
-                  "thingId": {
-                    "description": "The unique identifier for a thing",
-                    "type": "integer"
-                  }
-                },
-                "required": [ "thingId" ]
-              }
-            """,
-          )
-        }
-
-        cas2ApplicationEntityFactory.produceAndPersist {
-          withCrn(offenderDetails.otherIds.crn)
-          withId(applicationId)
-          withApplicationSchema(applicationSchema)
-          withCreatedByUser(submittingUser)
-        }
-
-        val resultBody = webTestClient.put()
-          .uri("/cas2/applications/$applicationId")
-          .header("Authorization", "Bearer $jwt")
-          .bodyValue(
-            UpdateCas2Application(
-              data = mapOf("thingId" to 123),
-              type = UpdateApplicationType.CAS2,
-            ),
-          )
-          .exchange()
-          .expectStatus()
-          .isOk
-          .returnResult(String::class.java)
-          .responseBody
-          .blockFirst()
-
-        val result = objectMapper.readValue(resultBody, Application::class.java)
-
-        assertThat(result.person.crn).isEqualTo(offenderDetails.otherIds.crn)
       }
     }
   }
@@ -1907,69 +1751,6 @@ class ApplicationTest : IntegrationTestBase() {
           val persistedAssessment = persistedApplication.getLatestAssessment() as TemporaryAccommodationAssessmentEntity
 
           assertThat(persistedAssessment.summaryData).isEqualTo("{\"num\":50,\"text\":\"Hello world!\"}")
-        }
-      }
-    }
-  }
-
-  @Test
-  fun `Submit Cas2 application returns 200`() {
-    `Given a User`(
-      staffUserDetailsConfigBlock = {
-        withTeams(
-          listOf(
-            StaffUserTeamMembershipFactory().produce(),
-          ),
-        )
-      },
-    ) { submittingUser, jwt ->
-      `Given a User` { userEntity, _ ->
-        `Given an Offender` { offenderDetails, inmateDetails ->
-          val applicationId = UUID.fromString("22ceda56-98b2-411d-91cc-ace0ab8be872")
-
-          val applicationSchema = cas2ApplicationJsonSchemaEntityFactory.produceAndPersist {
-            withAddedAt(OffsetDateTime.now())
-            withId(UUID.randomUUID())
-            withSchema(
-              """
-              {
-                "${"\$schema"}": "https://json-schema.org/draft/2020-12/schema",
-                "${"\$id"}": "https://example.com/product.schema.json",
-                "title": "Thing",
-                "description": "A thing",
-                "type": "object",
-                "properties": {},
-                "required": []
-              }
-            """,
-            )
-          }
-
-          cas2ApplicationEntityFactory.produceAndPersist {
-            withCrn(offenderDetails.otherIds.crn)
-            withId(applicationId)
-            withApplicationSchema(applicationSchema)
-            withCreatedByUser(submittingUser)
-            withData(
-              """
-              {}
-            """,
-            )
-          }
-
-          webTestClient.post()
-            .uri("/cas2/applications/$applicationId/submission")
-            .header("Authorization", "Bearer $jwt")
-            .header("X-Service-Name", ServiceName.cas2.value)
-            .bodyValue(
-              SubmitCas2Application(
-                translatedDocument = {},
-                type = "CAS2",
-              ),
-            )
-            .exchange()
-            .expectStatus()
-            .isOk
         }
       }
     }
